Delete old "Create" method
